Feature/login logic

+++ MrLocalBackend/Services/ProductService.cs
@@ -1,9 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using MrLocalBackend.Repositories.Interfaces;
+﻿using MrLocalBackend.Repositories.Interfaces;
 using MrLocalBackend.Services.Interfaces;
 using MrLocalDb.Entities;
+using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace MrLocalBackend.Services
 {
@@ -20,40 +20,63 @@ namespace MrLocalBackend.Services
             _validateData = validateData;
         }
 
-        public async Task<Product> AddProductToShop(string shopId, string name, string description, string priceType, decimal? price)
+        public async Task<Product> AddProductToShop(string shopId, string name, string description, string priceType, decimal? price, string userId)
         {
-            await _validateData.Value.ValidateProductData(shopId, name, description, (decimal)price, false, priceType);
-            var createdProduct = await _productRepository.Create(shopId, name, description, priceType, (decimal)price);
-            return createdProduct;
-        }
+            var shop = await _shopRepository.FindOne(shopId);
 
-        public async Task<Product> UpdateProduct(string id, string shopId, string name, string description, string priceType, decimal? price)
-        {
-            var product = await _productRepository.FindOne(id);
-
-            if (product == null)
+            if (shop.UserId == userId)
             {
-                throw new ArgumentException("Product to update doesn't exist");
+
+                await _validateData.Value.ValidateProductData(shopId, name, description, (decimal)price, false, priceType);
+                var createdProduct = await _productRepository.Create(shopId, name, description, priceType, (decimal)price);
+                return createdProduct;
             }
 
-            await _validateData.Value.ValidateProductData(shopId, name, description, price, true, priceType);
-            var updatedProduct = await _productRepository.Update(id, shopId, name, description, priceType, price);
-            return updatedProduct;
+            throw new ArgumentException("User cannot add products to this shop");
         }
 
-        public async Task<string> DeleteProduct(string id)
+        public async Task<Product> UpdateProduct(string id, string shopId, string name, string description, string priceType, decimal? price, string userId)
+        {
+            var shop = await _shopRepository.FindOne(shopId);
+
+            if (shop.UserId == userId)
+            {
+                var product = await _productRepository.FindOne(id);
+
+                if (product == null)
+                {
+                    throw new ArgumentException("Product to update doesn't exist");
+                }
+
+                await _validateData.Value.ValidateProductData(shopId, name, description, price, true, priceType);
+                var updatedProduct = await _productRepository.Update(id, shopId, name, description, priceType, price);
+                return updatedProduct;
+            }
+
+            throw new ArgumentException("User cannot add products to this shop");
+        }
+
+        public async Task<string> DeleteProduct(string id, string userId)
         {
             var products = await _productRepository.FindOne(id);
+            var shop = await _shopRepository.FindOne(products.ShopId);
 
-            if (products != null)
+            if (shop.UserId == userId)
             {
-                var deletedProduct = await _productRepository.Delete(id);
-                return deletedProduct;
+
+                if (products != null)
+                {
+                    var deletedProduct = await _productRepository.Delete(id);
+                    return deletedProduct;
+                }
+                else
+                {
+                    throw new ArgumentException("Invalid products parameters for deleting");
+                }
+
             }
-            else
-            {
-                throw new ArgumentException("Invalid products parameters for deleting");
-            }
+
+            throw new ArgumentException("User cannot delete products of this shop");
         }
 
         public async Task<List<Product>> GetAllProducts(string shopId)
